Added support for game resources and game modes

+++ ElectrodZMultiplayer/Core/Data/Messages/GameRestartRequestedMessageData.cs
@@ -16,14 +16,14 @@ namespace ElectrodZMultiplayer.Data.Messages
         /// Time to restart game in seconds
         /// </summary>
         [JsonProperty("time")]
-        public float Time { get; set; }
+        public double Time { get; set; }
 
         /// <summary>
         /// Is valid
         /// </summary>
         public override bool IsValid =>
             base.IsValid &&
-            (Time >= 0.0f);
+            (Time >= 0.0);
 
         /// <summary>
         /// Constructs a game restart request message for deserializers
@@ -37,9 +37,9 @@ namespace ElectrodZMultiplayer.Data.Messages
         /// Constructs a game restart request message
         /// </summary>
         /// <param name="time">Time to restart game in seconds</param>
-        public GameRestartRequestedMessageData(float time) : base(Naming.GetMessageTypeNameFromMessageDataType<GameRestartRequestedMessageData>())
+        public GameRestartRequestedMessageData(double time) : base(Naming.GetMessageTypeNameFromMessageDataType<GameRestartRequestedMessageData>())
         {
-            if (time < 0.0f)
+            if (time < 0.0)
             {
                 throw new ArgumentException("Time must be positive.", nameof(time));
             }
